♻️ extract bank trade formatting from /trade command
Move listing types and bank line helpers into tradeBank so the command only orchestrates fetch and reply.
